fix:The link of upload files is broken in the chat view #466
