+++ backend/app/api/main.py
@@ -7,7 +7,7 @@ from app.api.routes import (
     user,
     api_key,
     feedback,
-    download,
+    document,
 )
 from app.api.admin_routes.knowledge_base.routes import (
     router as admin_knowledge_base_router,
@@ -53,7 +53,7 @@ api_router.include_router(chat.router, tags=["chat"])
 api_router.include_router(feedback.router, tags=["chat"])
 api_router.include_router(user.router, tags=["user"])
 api_router.include_router(api_key.router, tags=["auth"])
-api_router.include_router(download.router, tags=["download_file"])
+api_router.include_router(document.router, tags=["documents"])
 api_router.include_router(admin_chat_engine.router, tags=["admin/chat_engine"])
 api_router.include_router(admin_document_router, tags=["admin/documents"])
 api_router.include_router(admin_feedback.router, tags=["admin/feedback"])
